requirements.txt
scipy==1.1.0
matplotlib==2.2.2
tensorflow==1.9.0
Werkzeug==0.14.1
Flask==1.0.2
numpy==1.14.5
pika==0.11.2
cfenv==0.5.3
Pillow==5.2.0
